pyproject.toml
[project]
name = "finance-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "dotenv>=0.9.9",
    "fastapi>=0.127.1",
    "fastapi-sso>=0.19.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.30",
    "uuid6>=2025.0.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mock-alchemy>=0.2.6",
    "mypy>=1.19.1",
    "pgmock>=1.3.7",
    "polyfactory>=3.2.0",
    "pytest-postgresql>=7.0.2",
    "ruff>=0.14.10",
    "testing-postgresql>=1.3.0",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
ignore = ["D100", "D101", "D102", "D103"]

[tool.mypy]
install_types=true
non_interactive=true
disable_error_code = ["abstract"]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_return_any = false
warn_unused_ignores = false
